move NAG optimizer function tests to separate module.
