+++ estimagic/tests/optimization/test_utilities.py
@@ -7,10 +7,6 @@ from estimagic.config import IS_DFOLS_INSTALLED
 from estimagic.config import IS_PETSC4PY_INSTALLED
 from estimagic.config import IS_PYBOBYQA_INSTALLED
 from estimagic.optimization import AVAILABLE_ALGORITHMS
-from estimagic.optimization.nag_optimizers import _change_evals_per_point_interface
-from estimagic.optimization.nag_optimizers import (
-    _perturb_jacobian_or_trust_region_step_from_user_value,
-)
 from estimagic.optimization.utilities import calculate_initial_trust_region_radius
 from estimagic.optimization.utilities import chol_params_to_lower_triangular_matrix
 from estimagic.optimization.utilities import cov_matrix_to_params
@@ -167,38 +163,3 @@ def test_available_algorithms():
     assert ("nag_dfols" in present_algo_names) is IS_DFOLS_INSTALLED
     assert ("tao_pounders" in present_algo_names) is IS_PETSC4PY_INSTALLED
     assert ("nag_pybobyqa" in present_algo_names) is IS_PYBOBYQA_INSTALLED
-
-
-def test_change_evals_per_point_interface_none():
-    res = _change_evals_per_point_interface(None)
-    assert res is None
-
-
-def test_change_evals_per_point_interface_func():
-    def return_args(trust_region_radius, min_trust_region, n_iterations, n_restarts):
-        return trust_region_radius, min_trust_region, n_iterations, n_restarts
-
-    func = _change_evals_per_point_interface(return_args)
-    res = func(delta=0, rho=1, iter=2, nrestarts=3)
-    expected = (0, 1, 2, 3)
-    assert res == expected
-
-
-def test_perturb_jacobian_or_trust_region_step_from_user_value_none():
-    res = _perturb_jacobian_or_trust_region_step_from_user_value(None)
-    assert res == (None, None)
-
-
-def test_perturb_jacobian_or_trust_region_step_from_user_value_jacobian():
-    res = _perturb_jacobian_or_trust_region_step_from_user_value("Jacobian")
-    assert res == (True, False)
-
-
-def test_perturb_jacobian_or_trust_region_step_from_user_value_trust():
-    res = _perturb_jacobian_or_trust_region_step_from_user_value("trust_region_step")
-    assert res == (False, True)
-
-
-def test_perturb_jacobian_or_trust_region_step_from_user_value_error():
-    with pytest.raises(ValueError):
-        _perturb_jacobian_or_trust_region_step_from_user_value("wrong_input")
